honeypot-ml/ml/requirements.txt
pymongo
river
scikit-learn
matplotlib
geoip2
python-dotenv
certifi 